requirements.txt
altair==5.2.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
branca==0.7.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
cryptography==42.0.5
DateTime==5.4
decorator==5.1.1
executing==2.0.1
folium==0.16.0
geojson==3.1.0
gitdb==4.0.11
GitPython==3.1.42
gmaps==0.9.0
gmplot==1.4.1
googlemaps==4.10.0
idna==3.6
importlib-metadata==7.0.1
ipython==8.22.1
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyterlab_widgets==3.0.10
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
numpy==1.26.4
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
pip-system-certs==4.0
prompt-toolkit==3.0.43
protobuf==4.25.3
pure-eval==0.2.2
pyarrow==15.0.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.17.2
pyOpenSSL==24.0.0
python-certifi-win32==1.6.1
python-dateutil==2.9.0
pytz==2024.1
referencing==0.33.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
setuptools-scm==8.0.4
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.31.1
streamlit-folium==0.18.0
streamlit-lottie==0.0.5
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.22.0
watchdog==4.0.0
wcwidth==0.2.13
widgetsnbextension==4.0.10
wrapt==1.16.0
xyzservices==2023.10.1
zipp==3.17.0
zope.interface==6.2
